Split generate-textbook-toc hook in utils to facilitate testing
